Program.cs -> App.xaml, AuthControl
